backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=43.0.1",
    "fastapi[standard]>=0.114.1",
    "pydantic-settings>=2.5.2",
    "pydantic>=2.8.2",
    "python-jose>=3.3.0",
    "requests>=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
hello = "backend:hello"

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.2",
    "fastapi-cli>=0.0.5",
]
